Mammoth 50K

Downloads the 50,000 point 'Mammoth' dataset, a 3D point cloud of a mammoth skeleton.

Usage

download_mammoth50k()

Format

A data frame with 50,000 rows and 3 variables, Y, X, Z, containing the Y, X, and Z coordinates respectively. This labeling of axes preserves the ordering of the raw data, and makes the Z-coordinate the "height", i.e. the height of the mammoth varies with Z. Note that the ordering of the axes differs from that of download_mammoth10k(). Use the X, Y, and Z labels rather than numerical indices for consistency between the two datasets.

For more information see https://pair-code.github.io/understanding-umap/.

Value

Data frame containing the Mammoth coordinates.

Details

Downloads a dataframe containing the 50,000 3D coordinates of a mammoth skeleton, digitized by The Smithsonian Institute.

This dataset is from Understanding UMAP, based on work originally done by Max Noichl. 50,000 points were down-sampled from the raw data used by Max Noichl.

Note

Requires the rjson package to be installed and loaded.

See also

Examples

if (FALSE) { # \dontrun{
mammoth <- download_mammoth50k()
# The mammoth in profile, facing left
plot(mammoth$X, mammoth$Z)
} # }