Two Linked Rings

Points sampled from two linked circles, from "How to Use t-SNE Effectively".

Usage

link_data(n)

Arguments

n

Number of points to sample.

Value

Data frame with coordinates in the x, y, z columns, and color in the color column.

Details

Creates a 3D dataset where points are uniformly sampled from two (2D) circles of unit radius, separated by 1 unit, and hence are overlapping. The two circles are not co-planar. Points are colored depending from which circle they are sampled from.

For the unlinked version, see unlink_data.

References

http://distill.pub/2016/misread-tsne/

See also

Examples

df <- link_data(n = 50)